Use MUnit in FrameDecoderSuite
